SettingValues: Track caches to invalidate.
And do not over-invalidate (if tag settings don't change, don't
invalidate tag cache), or under-invalidate (if votes change,
invalidate autosearch).

+++ src/settings/s_options.php
@@ -358,23 +358,21 @@ class Options_SettingParser extends SettingParser {
             $newj[] = $o->unparse();
         }
         $sv->save("next_optionid", null);
-        $sv->save("options", empty($newj) ? null : json_encode_db($newj));
-
-        $deleted_ids = array();
-        foreach ($sv->conf->options()->nonfixed() as $o) {
-            $newo = $this->stashed_options[$o->id] ?? null;
-            if (!$newo
-                || ($newo->type !== $o->type
-                    && !$newo->change_type($o, true, true)
-                    && !$o->change_type($newo, false, true))) {
-                $deleted_ids[] = $o->id;
+        if ($sv->update("options", empty($newj) ? null : json_encode_db($newj))) {
+            $deleted_ids = array();
+            foreach ($sv->conf->options()->nonfixed() as $o) {
+                $newo = $this->stashed_options[$o->id] ?? null;
+                if (!$newo
+                    || ($newo->type !== $o->type
+                        && !$newo->change_type($o, true, true)
+                        && !$o->change_type($newo, false, true))) {
+                    $deleted_ids[] = $o->id;
+                }
             }
+            if (!empty($deleted_ids)) {
+                $sv->conf->qe("delete from PaperOption where optionId?a", $deleted_ids);
+            }
+            $sv->mark_invalidate_caches(["options" => true, "autosearch" => true]);
         }
-        if (!empty($deleted_ids)) {
-            $sv->conf->qe("delete from PaperOption where optionId?a", $deleted_ids);
-        }
-
-        // invalidate cached option list
-        $sv->conf->invalidate_caches(["options" => true]);
     }
 }
